// packages/compiler/src/render3/partial/api.d.ts
export interface R3PartialDeclaration {
    /**
     * The minimum version of the compiler that can process this partial declaration.
     */
    minVersion: string;
    /**
     * Version number of the Angular compiler that was used to compile this declaration. The linker
     * will be able to detect which version a library is using and interpret its metadata accordingly.
     */
    version: string;
    /**
     * A reference to the `@angular/core` ES module, which allows access
     * to all Angular exports, including Ivy instructions.
     */
    ngImport: o.Expression;
    /**
     * Reference to the decorated class, which is subject to this partial declaration.
     */
    type: o.Expression;
}
// ...
export interface R3DeclareNgModuleMetadata extends R3PartialDeclaration {
    /**
     * An array of expressions representing the bootstrap components specified by the module.
     */
    bootstrap?: o.Expression[];
    /**
     * An array of expressions representing the directives and pipes declared by the module.
     */
    declarations?: o.Expression[];
    /**
     * An array of expressions representing the imports of the module.
     */
    imports?: o.Expression[];
    /**
     * An array of expressions representing the exports of the module.
     */
    exports?: o.Expression[];
    /**
     * The set of schemas that declare elements to be allowed in the NgModule.
     */
    schemas?: o.Expression[];
    /** Unique ID or expression representing the unique ID of an NgModule. */
    id?: o.Expression;
}